backend/e6_f5_s2_assessment/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
